# docs/openapi.yaml
openapi: 3.0.3
info:
  title: QUANLYCMS API
  version: "1.0.0"
  description: |
    API backend QUANLYCMS (Express + MongoDB).
    Auth: Bearer JWT. Mutating requests: CSRF cookie `csrf_token` + header `X-CSRF-Token`.
    Super Admin tenant scope: header `X-Tenant-Id`.
servers:
  - url: http://localhost:5000
    description: Local
  - url: /api
    description: Same-origin (Vite proxy / production)

tags:
  - name: Health
  - name: Auth
  - name: Notifications
  - name: Files
  - name: Backups
  - name: Monitoring
  - name: AI
  - name: BI
  - name: Workflows
  - name: Builder
  - name: Tenants

paths:
  /healthz:
    get:
      tags: [Health]
      summary: Public health probe
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  status: { type: string }
                  db: { type: string }
                  redis: { type: string }
                  queue: { type: string }
                  uptime: { type: number }
                  memory:
                    type: object
                    properties:
                      rssMb: { type: number }
                      heapUsedMb: { type: number }
        "503":
          description: Degraded (DB down)

  /api/auth/csrf-token:
    get:
      tags: [Auth]
      summary: Issue CSRF cookie + token
      responses:
        "200":
          description: CSRF token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  csrfToken: { type: string }

  /api/auth/login:
    post:
      tags: [Auth]
      summary: Admin / staff login (public portal may differ)
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Tokens or MFA challenge
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  mfaRequired: { type: boolean }
                  mfaToken: { type: string }
                  data:
                    type: object
                    properties:
                      accessToken: { type: string }
                      refreshToken: { type: string }

  /api/auth/mfa/verify:
    post:
      tags: [Auth]
      summary: Verify Super Admin TOTP after password
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [mfaToken, code]
              properties:
                mfaToken: { type: string }
                code: { type: string }
      responses:
        "200":
          description: Access tokens issued

  /api/notifications:
    get:
      tags: [Notifications]
      summary: List notifications (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: unreadOnly
          schema: { type: boolean }
      responses:
        "200":
          description: Notification list
    post:
      tags: [Notifications]
      summary: Admin broadcast
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "201":
          description: Created

  /api/notifications/count:
    get:
      tags: [Notifications]
      summary: Unread badge count
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Count

  /api/files/upload:
    post:
      tags: [Files]
      summary: Upload file by category
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
        - in: query
          name: category
          schema:
            type: string
            enum: [messages, assignments, training, invoices, images, general, popup, signature, logo]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: File registered

  /api/files:
    get:
      tags: [Files]
      summary: List file assets (admin)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: File list

  /api/backups:
    get:
      tags: [Backups]
      summary: List backup jobs (Super Admin)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Backup jobs
    post:
      tags: [Backups]
      summary: Queue a backup job
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "202":
          description: Queued

  /api/monitoring/overview:
    get:
      tags: [Monitoring]
      summary: Health + metrics + alerts
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Overview snapshot

  /api/ai/status:
    get:
      tags: [AI]
      summary: AI configuration status
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: configured / model

  /api/ai/quiz:
    post:
      tags: [AI]
      summary: Generate quiz questions
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                topic: { type: string }
                subject: { type: string }
                count: { type: integer }
      responses:
        "200":
          description: Questions (LLM or fallback)

  /api/bi/overview:
    get:
      tags: [BI]
      summary: Executive KPIs
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: period
          schema: { type: string, enum: [7d, 1m, 2m, 1y] }
        - in: query
          name: branchId
          schema: { type: string }
        - $ref: "#/components/parameters/TenantHeader"
      responses:
        "200":
          description: KPI payload

  /api/workflows:
    get:
      tags: [Workflows]
      summary: List workflow instances
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: sync
          schema: { type: boolean }
      responses:
        "200":
          description: Instances

  /api/workflows/{id}/advance:
    post:
      tags: [Workflows]
      summary: Approve or reject workflow
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [approve, reject] }
                note: { type: string }
      responses:
        "200":
          description: Updated instance

  /api/builder/forms:
    get:
      tags: [Builder]
      summary: List forms
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Forms
    post:
      tags: [Builder]
      summary: Create form
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "201":
          description: Created

  /api/builder/reports:
    get:
      tags: [Builder]
      summary: List report definitions
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Reports

  /api/builder/reports/{id}/run:
    get:
      tags: [Builder]
      summary: Run report
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Rows + columns

  /api/tenants:
    get:
      tags: [Tenants]
      summary: List tenants (Super Admin)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Tenants
    post:
      tags: [Tenants]
      summary: Create tenant
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "201":
          description: Created

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CsrfHeader:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string }
      description: Must match csrf_token cookie
    TenantHeader:
      in: header
      name: X-Tenant-Id
      required: false
      schema: { type: string }
      description: Super Admin tenant scope